This builds and runs, but sometimes gets Empty Leaf asserts
Fix upstream issue with C4305 in lltraceaccumulators.h, use constexpr more.
Adds dullahan, updates a bunch of libraries.
We use LLCoreHTTP now.
OpenJPEG is now built out of tree, that thing was getting out of date.
The brunt of this merge was done back in August, June, and May last year.

Stab at LLAlertDialog, we should really just update to LLScreenChannel...
There may be a double free hiding in the notifications shared pointers here
I couldn't track down why, though.

+++ indra/llcommon/llstreamtools.cpp
@@ -1,3 +1,5 @@
+// This is an open source non-commercial project. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++ and C#: http://www.viva64.com
 /** 
  * @file llstreamtools.cpp
  * @brief some helper functions for parsing legacy simstate and asset files.
@@ -118,7 +120,7 @@ bool skip_to_next_word(std::istream& input_stream)
 
 bool skip_to_end_of_next_keyword(const char* keyword, std::istream& input_stream)
 {
-	int key_length = strlen(keyword);	 /*Flawfinder: ignore*/
+	size_t key_length = strlen(keyword);	 /*Flawfinder: ignore*/
 	if (0 == key_length)
 	{
 		return false;
@@ -315,7 +317,7 @@ bool unget_line(const std::string& line, std::istream& input_stream)
 // returns true if removed last char
 bool remove_last_char(char c, std::string& line)
 {
-	int line_size = line.size();
+	size_t line_size = line.size();
 	if (line_size > 1
 		&& c == line[line_size - 1])
 	{
